tools/make-repo.sh: add a tool to generate a local repo with osbuild rpms
Run the tool as `./make-repo.sh fedora 30 x86_64`.

Tihs generates rpms for osbuild and osbuild-composer, as they are
in current HEAD of the current repo (and its submodules).

A repo is made from the rpms and served on http://localhost:8000,
when the web server is terminated the repository is cleand up.

The sha256 checksum of the repository is printed to the console, this
is needed if using the repo in a pipeline.

Additionally, a pipeline test.json is added, which can be used to generate
a bootable image running cockpit with osbuild-composer as backend. The
pipeline must be initialized with the checksum printed by make-repo.sh
and osbuild from the submodule must be run on this pipeline while the
local repo is still running.

Boot the image and connect to its IP address on port 9090 to access the
cockpit web-interface. The user/pass is test/password.

Signed-off-by: Tom Gundersen <teg@jklm.no>