distro: use fedora-30 as hostname for all fedora versions
The only fedora image that can be built currently is for fedora 30.
Other versions of fedora can successfully build fedora 30 images but
when the distro name/version is checked it is not recognized as a valid
distro so osbuild-composer will not run. All fedora versions now use
the hostname/version of fedora-30 regardless if they are actually fedora
30.