test/cases: merge remaining empty_blueprint and local_boot
As in 616b6250c7fed010358e34f9fb9dbfbed3ada259, add the needed
ssh key to the formerly empty blueprint, and use this test-case
for booting as well as pipeline generation verification.

For the ext4-partition image type, we also needed to add the
openssh-server package, as @core is not included by default.

Signed-off-by: Tom Gundersen <teg@jklm.no>